+++ CHANGELOG.md
@@ -445,3 +445,4 @@
 * Define `Prism` type and its `preview` method
 * Redefine `Lens` type as `Optics` on `Identity`
 * Define `#=@` type operator as synonymous for `Optics`
+* [WIP] Redefine `Substructure` typeclass with polymorphic `Optics`